---ajuste al proceso para que lo modelos de negocio contable y flujo provengan de un mismo CC. ---Se ajustan tipos de int a int64 donde fue necesario ---se elimina clase inecsaria

+++ AppGia/Controllers/ProformaDataAccessLayer.cs
@@ -579,7 +579,7 @@ namespace AppGia.Controllers
             }
         }
 
-        public List<ProformaDetalle> CalculaDetalleProforma(Int64 idCenCos, int mesInicio, int idEmpresa, Int64 idModeloNeg, int idProyecto, int anio, Int64 idTipoCaptura,Int64 idTipoProforma)
+        public List<ProformaDetalle> CalculaDetalleProforma(Int64 idCenCos, int mesInicio, Int64 idEmpresa, Int64 idModeloNeg, Int64 idProyecto, int anio, Int64 idTipoCaptura,Int64 idTipoProforma)
         {
             ///obtener las variables
             ProformaDetalleDataAccessLayer objProfDetalle = new ProformaDetalleDataAccessLayer();
